backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
structlog==23.2.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
aiofiles==23.2.1
